--- routes/main.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-router.get('/', (req, res) =>{
-    res.send('This is main page');
-});
-
-module.exports = router;
