+++ routes/mypageRouter.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const mypageRouter = express.Router();
+
+mypageRouter.get('/', (req, res) =>{
+    res.send('This is "MYPAGE" page');
+});
+
+module.exports = mypageRouter;
